Guards, enhancements and search for users and posts

+++ src/api/posts/posts.service.ts
@@ -1,9 +1,10 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
 import { Post } from './entities/post.entity';
+import e from 'express';
 
 @Injectable()
 export class PostsService {
@@ -21,28 +22,51 @@ export class PostsService {
     return { posts, total };
   }
 
-  async createPost(createPostDto: CreatePostDto): Promise<Post> {
-    const newPost = this.postRepository.create(createPostDto);
-    return this.postRepository.save(newPost);
+  async createPost(createPostDto: CreatePostDto, userReq: any): Promise<Post> {
+    if (userReq.id == createPostDto.author_id || userReq.role == 'admin') {
+      const newPost = this.postRepository.create(createPostDto);
+      return this.postRepository.save(newPost);
+    } else {
+      throw new UnauthorizedException('You are not authorized to perform this action');
+    }
   }
 
-  async updatePost(id: string, updatePostDto: UpdatePostDto): Promise<Post> {
+  async updatePost(id: string, updatePostDto: UpdatePostDto, userReq: any): Promise<Post> {
     const post = await this.postRepository.findOne({ where: { id: +id } });
     if (!post) {
       throw new NotFoundException(`Post with ID ${id} not found`);
     }
 
-    Object.assign(post, updatePostDto);
-    return this.postRepository.save(post);
+    if (userReq.id == post.author_id || userReq.role == 'admin') {
+      const post = await this.postRepository.findOne({ where: { id: +id } });
+      if (!post) {
+        throw new NotFoundException(`Post with ID ${id} not found`);
+      }
+
+      Object.assign(post, updatePostDto);
+      return this.postRepository.save(post);
+    } else {
+      throw new UnauthorizedException('You are not authorized to perform this action');
+    }
   }
 
-  async deletePost(id: string): Promise<{ message: string }> {
+  async deletePost(id: string, userReq: any): Promise<{ message: string }> {
     const post = await this.postRepository.findOne({ where: { id: +id } });
     if (!post) {
       throw new NotFoundException(`Post with ID ${id} not found`);
     }
 
-    await this.postRepository.remove(post);
-    return { message: 'Post deleted successfully' };
+    if (userReq.id == post.author_id || userReq.role == 'admin') {
+      await this.postRepository.remove(post);
+      return { message: 'Post deleted successfully' };
+    } else {
+      throw new UnauthorizedException('You are not authorized to perform this action');
+    }
+  }
+
+  async searchPosts(query: string): Promise<Post[]> {
+    return this.postRepository.createQueryBuilder('post')
+        .where('post.title LIKE :query OR post.content LIKE :query', { query: `%${query}%` })
+        .getMany();
   }
 }
